#3 - `KalturaAPI` method signatures are different
Maybe it's a little evil of me to do this.  `KalturaAPI` methods didn't need all the same arguments as the `MiVideoAPI` ones, so I made some of the arguments optional with a default of `NotImplemented`.  Then I moved the required arguments to the beginning.  That made calls in `KalturaCaptionLoader` not work correctly with `KalturaAPI`.  Specifying the argument names here solves that problem.

86 47